requirements.txt
alembic==1.0.8
alpaca-trade-api==0.26
APScheduler==3.6.0
asyncio-nats-client==0.8.2
certifi==2018.11.29
chardet==3.0.4
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.4.0
Flask-Moment==0.7.0
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.1
nltk==3.4
numpy==1.16.1
pandas==0.24.1
pkg-resources==0.0.0
praw==6.1.1
prawcore==1.0.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2018.9
redis==3.2.1
requests==2.21.0
rq==0.13.0
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.3.0
tzlocal==1.5.1
update-checker==0.16
urllib3==1.24.1
visitor==0.1.3
websocket-client==0.54.0
websockets==7.0
Werkzeug==0.14.1
